web_service/requirements.txt
Flask
Flask-Session
python-dotenv
redis
gunicorn
bcrypt
pyjwt
uuid